# api/openapi-server-v1.yaml
openapi: 3.1.0
info:
  title: aimee-server API
  version: "1"
  description: >
    Native HTTP surface for aimee-server. Exposes liveness/discovery
    endpoints, an OpenAI-compatible shaping layer
    (models / chat.completions / completions / embeddings), and persona
    management. The same operations are also available over the NDJSON
    socket; this `/v1` surface is the contract-first HTTP transport.
servers:
  - url: http://127.0.0.1:{port}/v1
    variables:
      port:
        default: "8910"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Required for `/v1/*` over TCP. Unix-domain-socket access is
        authorised by filesystem permissions and needs no token.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
...
paths:
...
  /rules:
    get:
      summary: List active collaboration rules
      responses:
        "200":
          description: Active collaboration rules ({epoch, rules[]})
          content:
            application/json:
              schema:
                type: object
                properties:
                  epoch: { type: integer }
                  rules: { type: array, items: { type: object } }
        "502":
          description: Rules backend (aimee-kb) unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "503":
          description: Rules not available on this server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /workflow/blocks/{name}:
    put:
      summary: Create or edit a delegate custom block (blocks.yaml; command executors refused)
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Updated block catalog, content: { application/json: { schema: { type: object } } } } }
    delete:
      summary: Delete a custom delegate block (command blocks are operator-only)
      responses: { "200": { description: Updated block catalog, content: { application/json: { schema: { type: object } } } } }
...
  /workflow/defs/{name}:
    get:
      summary: One workflow definition (canonical form, version, node graph)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "{valid,name,version,canonical,def,error?}"
          content:
            application/json:
              schema: { type: object }
        "400":
          description: Invalid workflow name
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404":
          description: Workflow not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /workflow/validate:
    post:
      summary: Validate posted workflow YAML without saving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [yaml]
              properties:
                yaml: { type: string }
      responses:
        "200":
          description: "{valid,name,version,canonical,def,error?}"
          content:
            application/json:
              schema: { type: object }
        "400":
          description: Malformed request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /workflow/save:
    post:
      summary: Canonical-normalize + save a workflow (optimistic lock on prev_version)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, yaml]
              properties:
                name: { type: string }
                yaml: { type: string }
                prev_version: { type: string }
      responses:
        "200":
          description: "{name,version}"
          content:
            application/json:
              schema: { type: object }
        "400":
          description: Invalid definition or name
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "409":
          description: Version conflict / already exists
          content:
            application/json:
              schema: { type: object }
...
  /workflow/items/{id}:
    get:
      summary: One work item's run-state
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Work-item run-state
          content:
            application/json:
              schema: { type: object }
        "404":
          description: Work item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      summary: Delete a run after auto-abandoning it, then remove its rows and proposal file (owner or operator)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "{id, deleted:true}"
          content:
            application/json:
              schema: { type: object }
        "403":
          description: Not the caller's work item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404":
          description: Work item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /dev/submit:
    post:
      summary: Submit a proposal for autonomous development (runs server-side via the workflow engine)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proposal_md]
              properties:
                proposal_md: { type: string, description: The proposal to execute autonomously }
                workflow: { type: string, description: Workflow name (default "build") }
                repo: { type: string, description: Optional repo identifier }
      responses:
        "200":
          description: "{work_item_id, workflow, state}"
          content:
            application/json:
              schema: { type: object }
        "400":
          description: Missing proposal_md
        "500":
          description: Could not create the work item
...
  /runs:
    post:
      summary: Create and run an inference run (synchronous); retrievable by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                model: { type: string }
                input: {}
                temperature: { type: number }
                max_tokens: { type: integer }
      responses:
        "200":
          description: Completed run object
          content:
            application/json:
              schema: { type: object }
        "400":
          description: Invalid run request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "502":
          description: Inference failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "503":
          description: Runs not available on this server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /runs/{id}:
    get:
      summary: Retrieve a previously created run
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The run object
          content:
            application/json:
              schema: { type: object }
        "404":
          description: No such run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "410":
          description: Run interrupted by server replacement or evicted from the live store
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /agents:
    get:
      summary: List configured models and the default agent
      responses:
        "200":
          description: Configured agents ({default, agents[]})
          content:
            application/json:
              schema:
                type: object
                properties:
                  default: { type: string }
                  agents:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        provider: { type: string }
                        model: { type: string }
                        enabled: { type: boolean }
                        roles: { type: array, items: { type: string } }
        "502":
          description: Agent configuration unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "503":
          description: Agents not available on this server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /config/set:
    post:
      summary: Set one configuration value
      description: >
        Roundtable policy keys are operator-controlled and require the
        server-token-attested `webuser:admin` appliance administrator; generic
        capabilities and local Unix-socket access are insufficient.
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "200":
          description: Updated config key/value
          content:
            application/json:
              schema: { type: object }
        "403":
          description: Roundtable policy mutation was not authorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /cron/run:
    post:
      summary: Run a cron job now
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id]
              properties:
                job_id: { type: string }
      responses:
        "200":
          description: Cron run result
          content:
            application/json:
              schema: { type: object }
...
  /delegate/run:
    post:
      summary: Run a delegate task (forced background when invoked remotely; returns job_id to poll)
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Delegate run accepted, content: { application/json: { schema: { type: object } } } } }
...
  /vault/unlock:
    post:
      summary: "Unlock the credential vault over a kernel-attested local or webchat UDS connection"
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Unlocked, content: { application/json: { schema: { type: object } } } } }
...
  /model/roles:
    post:
      summary: Set an agent's roles
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Agent roles updated, content: { application/json: { schema: { type: object } } } } }
...
  /init/run:
    post:
      summary: Run project initialization
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: Init result, content: { application/json: { schema: { type: object } } } } }
...
security:
  - bearerAuth: []